recipes-shift: Add recipe to run graphical app before /sbin/init
One way to boot fast is to start the user interface first. As soon as
userspace is started, run the graphical interface, in this case
glmark2. Then the init scripts can continue running to bring up the
rest of the system.
If the GPU driver is built into the kernel, itwill be avaiable by
the time the kernel starts the userspace init process. This trick
starts the display 10s-20s sooner than the regular init.