recipes-kernel: Add patches for enabling display on DK2 boards
The HDMI bridge on DK2 boards is not initialized correctly by linux.
Even if the HDMI transmitter is not used, an error in its
initialization causes the display controller driver not to load. This
also takes out the DSI display DRM node, disabling the display panel.

The u-boot SSBL contained a workaround for this by enabling the
"v1v2_hdmi" and "v3v3_hdmi" supplies. In falcon mode, this code does
not run, so the kernel needs to correctly set the supply voltages.