recipes-bsp: Add recipe to assemble a FIT image with OPTEE and linux
In order to use falcon mode and skip u-boot, we need to load OPTEE
before jumping to the kernel. This can be achieved by instructing SPL
to load the required components via a FIT image.

The FIT image, may be used instead of the uImage.

FIT images support cryptographic signatures, but this is beyond the
scope of this change.