recipes-security: optee-os: Fix OPTEE_TZDRAM_START parsing
The "OPTEE_TZDRAM_START=" flag to the OP-TEE makefile was being set as
'TZDRAM_FLAGS' in do_configure(). However, this variable is not
passed to other stages, such as do_compile. Thus the TZDRAM_FLAGS were
not picked up during the build.

To resolve this, don't use do_configure for this, but instead use an
__anonymous python function. When set in this manned, the datastore is
preserved, and TZDRAM_FLAGS is correctly populated.

OP-TEE checks that the TZC configuration is correct, using both
CFG_TZDRAM_START, and CFG_DRAM_SIZE. Thus, these configs need to be in
sync. This change also infers DRAM size based on TZDRAM start address.