recipes-kernel: devicetree-overlays: Add recipe to compile .dto
The vanilla devicetree provided by linux is adjusted on boot to add
kernel boot arguments and other platform-specific information. Two
mechanisms exist, the 'spl export' command, and devicetree overlays.
We prefer devicetree overlays because they can be signed. Although
the kernel devicetree is not a secure resource, we prefer a boot flow
in which it can be verified.