recipes-bsp: u-boot: Enable devicetree overlays in SPL
In u-boot's falcon mode, a devicetree is modified by u-boot, then
written to persistent storage. The modified devicetree is used by
SPL as-is to bypass u-boot and load a Linux kernel directly. The
disadvantage of this approach is that the devicetree is not signed,
and cannot be verified at boot.

An alternative method is to boot a FIT, which can have all its
components signed, including the devicetree. The DTBs provided by
linux need some modification, for example, adding the boot arguments.
We choose to do this by applying one or more overlays to the linux
DTB. The overlays are also signed, allowing the modified devicetree to
be completely verified at boot time.

In preparation for switching the boot flow to include verified
devicetree images, enable DTB overlays in SPL.