conf/machine: Set the OP-TEE TZDRAM_START for -DK2 board
Ideally, OP-TEE would autodetect memory layout. However, some parts of
op-tee are hardcoded at compile time. This means we are not be able to
use the same image for EV1 and DK2 boards.

Of course supporting both boards is a more complicated endeavor, so
set up OPTEE_TZDRAM_START to match the memory layout on DK2 boards.