recipes-kernel: Make sure REGULATOR_FIXED_VOLTAGE is enabled
CONFIG_REGULATOR_FIXED_VOLTAGE is required for DK2 and EV1 boards
because the STPMIC is supplied by a fixed regulator. When this is not
enabled, stpmic probing returns -EPROBE_DEFER, and several
peripherals, including SD/MMC, will not work.

Fixed regulator support usually ends up enabled, but this is not
guaranteed. multi_v7_defconfig does not enable it, nor do any of the
config fragments. Explicitly enable fixed regulator support.