recipes-bsp: fit-image: Factor out 'mkimage' step into shell function
The process of creating the FIT image can be broken down to two steps.
  1. Creation of a .its file
  2. Running mkimage to assemble the final binary

On this basis alone, it makes sense to separate the steps into their
respective functions. Aditionally, the first step is easier to
implement in python, while the mkimage invocation is best suited for
shell code. Place the invocation of 'uboot-mkimage' into a separate
shell function.