recipes-bsp: fit-image: Make dependency on do_deploy() explicit
The kernel zImage and devicetree files are installed during the
kernel's do_deploy to ${DEPLOY_DIR_IMAGE}. They are not installed to
the recipe sysroot. do_deploy() is not guaranteed to run before
do_assemble_fitimage(). We might not have all the files we need.

This requires adding an explicit dependency on do_deploy().