recipes-bsp: fit-image: Workaround FDT load address limitations
The current version of u-boot does not understand where the FDT should
be loaded. It won't try to relocate the FDT, and instead modify it in
place. This works with only one FDT, but breaks down when overlays are
to be applied.

As a workaround, force the DTB and overlays to be loaded at specific
addresses until u-boot is fixed.