recipes-bsp: fit-image: Remove hardcoded FDT "load =" address
The "spl export" command does not figure out where to load overlay
nodes. It requires an explicit "load =" for each overlay. On the other
hand, SPL will find an appropriate load address on its own. Since we
are now relying on this new behavior, the "load =" properties are no
longer required. Remove them.