.gitlab-ci.yml: Enable continuous integration with Gitlab
We want a way to test meta-horseshift in isolation, not as a
submodule of horseshift-linux. This repository is itself a submodule,
so we can't add other submodules to it.

While unorthodox, the required repos are cloned during the build. The
gitlab cache ensures theya re only cloned once per pipeline.