recipes-bsp: u-boot: Update to v2021.04
Unfortunately, v2021.04 did not land all the bells and whistles needed
by our boot flow. However, some updates to the SPL boot bath did make
it, and we can drop one patch.

Unfortunately, one of the changes that landed in v2021.04 breaks our
boot flow, so we have to revert it.