recipes-kernel: overlays: Add overlay for OP-TEE support on DK2
In an optimal flow, u-boot SPL would add the optee nodes to the
devicetree on boot. However, this support is not yet available
upstream. An alternative is to include the optee nodes as an overlay.

This is a temporary workaround until u-boot upstream catches up. It is
currently only implemented for DK2, as EV1 has different reserved
memory regions. The expectation is that a future u-boot will
automatically take care of the differences, depending on the board.