.gitlab-ci.yml: Share cache across the same branch
Originally Gitlab caches were set up to be unique to each pipeline.
This setup provided a completely clean build, each time, but at the
expense of downloading sources each time. If one download is too
slow, as was recently the case with u-boot, the entire pipeline can
timeout.

A more efficient was is to share the cache across the same branch. So
update the cache key to refernce the branch name, not the pipeline ID.