recipes-bsp: u-boot: Explicitly overwrite the SRC_URI
Upstream OpenEmbedded has decided it's a good idea to willie-nillie
add patches to u-boot, overriding the choices of individual platform
maintainers. This could be as innocent as causing conflicts with our
patches to the build failing because the patches no longer apply.

This is easily resolved by setting SRC_URI instead of appending to it.
Since the patch names are in SRC_URI, by setting it we overrule the
common u-boot patches.

Note, we also point to source.denx.de, as the new home of the u-boot
sources.