recipes-shift: fastinit: Update license to "MIT" from "GPLv2+"
With the OpenEmbdedded 'hardknott' release, the "GPL-2.0" license
file is no longer available. It would have to be either "GPL-2.0-only"
or "GPL-2.0-or-later". However those files are not available in the
'dunfell' and 'gatesgarth' releases.

While we could manually add license files, the fact remains that
fastinit is simple and trivial enough that it doesn't really benefit
from being GPL. Instead, use MIT, which is the simplest license.