conf/distro: HorseSHIfT: Fix GCC-11 builds by disabling mklibs
The 'mklibs-native' package fails to build with GCC-11. This is
problematic on newer distros which have moved away from	GCC-10. On
these system the HoseSHIfT build fails because of mklibs.

The mklibs package is not needed, and is likely to be removed from the
next openembedded release anyway. For now, disable it.