recipes-bsp: u-boot: Disable the "fdtargs" partition
When CONFIG_SYS_MMCSD_RAW_MODE_ARGS_SECTOR is defined, SPL will check
for a devicetree at this sector. In our setup, this would be the
"fdtargs" partition, which is no longer used. Even if SPL won't use
the data, it still wastes time reading the partition. This is not
optimal.

Do not define CONFIG_SYS_MMCSD_RAW_MODE_ARGS_SECTOR, which causes SPL
to not try to read it. The partition can now be completely eliminated.