conf + recipes: Switch to new bitbake override syntax
In newer bitbake releases, the old VAR_append/remove syntax is being
replaced with a VAR:append/remove style. Bitbake versions above 1.50
will no longer accept the older style syntax. Convert configuration
files and recipes to this new syntax.

We also need to update the bitbake version to 1.50.3 for this new
syntax to be properly recognized.